HLE: Convert a lot more logging to the "automatic" parameter logging system

+++ ext/imgui/imgui_impl_thin3d.cpp
@@ -115,7 +115,7 @@ void ImGui_ImplThin3d_RenderDrawData(ImDrawData* draw_data, Draw::DrawContext *d
 			} else {
 				size_t index = (size_t)pcmd->TextureId - TEX_ID_OFFSET;
 				if (index >= bd->tempTextures.size()) {
-					WARN_LOG(Log::System, "Missing temp texture %d (out of %d)", index, (int)bd->tempTextures.size());
+					WARN_LOG(Log::System, "Missing temp texture %d (out of %d)", (int)index, (int)bd->tempTextures.size());
 					continue;
 				}
 				_dbg_assert_(index < bd->tempTextures.size());
